mlops/services/api/requirements.txt
fastapi==0.115.6
uvicorn[standard]==0.34.0
ultralytics==8.3.0
pillow==10.4.0
prometheus-client==0.20.0
python-multipart==0.0.9
